Created product search page, cleaned up
