+++ backend/api/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 
-# Create your models here.
+# NOTE: User model is already defined in django.contrib.auth.models
+    
 class Products(models.Model):
     name = models.CharField(max_length=100)
     description = models.TextField()
