+++ backend/api/serializers.py
@@ -10,4 +10,4 @@ class ProductSerializer(serializers.ModelSerializer):
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
-        fields = ['id', 'email', 'date_joined']
+        fields = ['id', 'email', 'password', 'date_joined']
