+++ frontend/src/pages/home.js
@@ -1,46 +1,9 @@
-import { useEffect, useState } from 'react';
-import axios from 'axios';
-
 const Home = () => {
-    // Use axios to get products from "http://127.0.0.1:8000/api/products"
-    const [products, setProducts] = useState([]);
-    useEffect(() => {
-        const fetchData = async () => {
-            const { data } = await axios.get('/api/products');
-            setProducts(data);
-        };
-        fetchData();
-    }, []);
     return (
         <div>
             <h1>Home</h1>
             <p1> Welcome to the home page</p1>
             <p2> there really isn't much here</p2>
-            <table>
-                <thead>
-                    <tr>
-                        <th style = {{ padding: '10px', textAlign: 'center' }}>ID</th>
-                        <th style = {{ padding: '10px', textAlign: 'center' }}>Name</th>
-                        <th style = {{ padding: '10px', textAlign: 'center' }}>Description</th>
-                        <th style = {{ padding: '10px', textAlign: 'center' }}>Category</th>
-                        <th style = {{ padding: '10px', textAlign: 'center' }}>Sustainability Factor</th>
-                        <th style = {{ padding: '10px', textAlign: 'center' }}>Image</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {/* I want more space in between columns */}
-                    {products.map((product) => (
-                        <tr key={product.id}>
-                            <td style={{ padding: '10px', textAlign: 'center' }}>{product.id}</td>
-                            <td style={{ padding: '10px', textAlign: 'center' }}>{product.name}</td>
-                            <td style={{ padding: '10px', textAlign: 'center' }}>{product.description}</td>
-                            <td style={{ padding: '10px', textAlign: 'center' }}>{product.category}</td>
-                            <td style={{ padding: '10px', textAlign: 'center' }}>{product.sustainability_factor}</td>
-                            <td style={{ padding: '10px', textAlign: 'center' }}><img src={product.image_link} alt={product.name} width="100" /></td>
-                        </tr>
-                    ))}
-                </tbody>
-            </table>
         </div>
     );
 }
